requirements.txt
flask
Werkzeug
torchvision
pillow
